Fix: Corrected the return types of certain system register helper functions.
   modified following function prototypes
   uint64_t bsa_gic_get_far(void)
   uint64_t bsa_gic_get_esr(void)
   void bsa_gic_update_elr(uint64_t elr_value)
   uint64_t bsa_gic_get_elr(void)
   uint64_t bsa_gic_ack_intr(void)

+++ val/driver/gic/gic.c
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2021,2023-2025, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2021,2023-2026, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -84,7 +84,7 @@ val_gic_acknowledgeInterrupt(void)
 
   gic_version = val_gic_get_info(GIC_INFO_VERSION);
   if ((gic_version == 3) || (gic_version == 4))
-      return v3_AcknowledgeInterrupt();
+      return (uint32_t) v3_AcknowledgeInterrupt();
   else
       return v2_AcknowledgeInterrupt();
 }
